INDY-1896: tests refactoring
